+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.1 (TBA)
 
 * Relax mint requirement
+* Fix bug in `Assent.HTTPAdapter.Mint` where HTTP/2 responses wasn't parsed correctly
 
 ## v0.1.0 (2019-10-06)